+++ src/error.rs
@@ -46,7 +46,7 @@ pub enum Error {
 
     #[cfg(feature = "parse-cfg")]
     /// Pass-thru `toml::de::Error`.
-    #[error("Serde_json Error: {0}")]
+    #[error("Serde Toml Error: {0}")]
     SerdeTomlError(#[from] toml::de::Error),
 
     #[cfg(feature = "matrix")]